Penambahan layout login, dan pembaruan controller admin role

+++ src/main/java/com/example/carikado/repository/GiftInfoRepository.java
@@ -10,6 +10,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface GiftInfoRepository extends JpaRepository<GiftInfo, Integer> {
+public interface GiftInfoRepository extends JpaRepository<GiftInfo, Long> {
 
 }
