+++ src/main/java/com/example/carikado/service/GiftInfoService.java
@@ -1,15 +1,12 @@
 package com.example.carikado.service;
 
 import com.example.carikado.model.GiftInfo;
-import com.example.carikado.model.GiftInfoCategory;
-import com.example.carikado.model.User;
 import com.example.carikado.repository.GiftInfoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @Service("giftInfoService")
@@ -34,7 +31,7 @@ public class GiftInfoService {
         return mGiftInfoRepository.findAll(pageable);
     }
 
-    public GiftInfo findGift(Integer id) {
+    public GiftInfo findGift(Long id) {
         return mGiftInfoRepository.findOne(id);
     }
 
@@ -42,7 +39,7 @@ public class GiftInfoService {
         return mGiftInfoRepository.save(giftInfo);
     }
 
-    public void deleteGiftInfo(Integer id) {
+    public void deleteGiftInfo(Long id) {
         mGiftInfoRepository.delete(id);
     }
 }
